# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking Reservation API",
        "description": "API for managing parking spot reservations, including booking, cancellation, and retrieving user reservations.",
        "termsOfService": "http://example.com/terms",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations/{parkingId}": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Reserve a parking spot",
                "description": "Allows a user to reserve a parking spot by providing their user ID and reservation timings.",
                "operationId": "aab918522397b8add755abb4f50054a9",
                "parameters": [
                    {
                        "name": "parkingId",
                        "in": "path",
                        "description": "ID of the parking spot to reserve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "arrival_time",
                                    "departure_time"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-03-14T08:00:00"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-03-14T10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "parking_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "arrival_time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "departure_time": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/reservations/{reservationId}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancel a reservation",
                "description": "Allows a user to cancel their reservation by ID.",
                "operationId": "e0a7332ac924a9ed7463c11ff1fc7418",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "description": "ID of the reservation to cancel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation cancelled successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/reservations/user/{userId}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get user reservations",
                "description": "Fetch all reservations of a specific user.",
                "operationId": "a768431d9f3ef1d5e7ccab7db61a7374",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reservations"
                    },
                    "404": {
                        "description": "No reservations found"
                    }
                }
            }
        },
        "/api/reservations/my": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get my reservations",
                "description": "Fetch all reservations of the currently authenticated user.",
                "operationId": "d8e56659e8c4c2efd81f16997b81d7e5",
                "responses": {
                    "200": {
                        "description": "List of reservations"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}